automation/requirements.txt
# Dependências existentes
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
webdriver-manager==4.0.1
pandas==2.1.3
numpy==1.24.3
openpyxl==3.1.2
python-dotenv==1.0.0

# Novas dependências para comparação de imagens e conversão de PDF
opencv-python==4.8.1.78
scikit-image==0.21.0
pdf2image==1.16.3
Pillow==10.0.1
Pillow-SIMD==9.0.0.post3  # Versão otimizada para performance

# Dependências adicionais para processamento de imagens
matplotlib==3.7.2
seaborn==0.12.2
